修改 ECS 序列化相关

+++ modules/runtime/include/binary/blob_fwd.h
@@ -6,24 +6,24 @@
 struct RUNTIME_API skr_blob_arena_t
 {
     skr_blob_arena_t();
-    skr_blob_arena_t(void* buffer, uint64_t base, size_t size, size_t align);
+    skr_blob_arena_t(void* buffer, uint64_t base, uint32_t size, uint32_t align);
     skr_blob_arena_t(skr_blob_arena_t&& other);
     skr_blob_arena_t& operator=(skr_blob_arena_t&& other);
 
     ~skr_blob_arena_t();
     void* get_buffer() const { return buffer; }
-    size_t get_size() const { return offset; }
-    size_t get_align() const { return align; }
+    uint32_t get_size() const { return offset; }
+    uint32_t get_align() const { return align; }
 #ifdef SKR_BLOB_ARENA_CHECK
-    void release(size_t);
+    void release(uint32_t);
 #endif
     uint64_t base() const { return _base; }
 private:
     void* buffer;
     uint64_t _base;
-    size_t align;
-    size_t offset;
-    size_t capacity;
+    uint32_t align;
+    uint32_t offset;
+    uint32_t capacity;
 };
 
 struct RUNTIME_API skr_blob_arena_builder_t
